Move some tests from core module to their respective module
