+++ app/src/androidTest/java/de/test/antennapod/ui/PreferencesTest.java
@@ -10,12 +10,12 @@ import androidx.test.filters.LargeTest;
 import androidx.test.rule.ActivityTestRule;
 import de.danoeh.antennapod.R;
 import de.danoeh.antennapod.ui.screen.preferences.PreferenceActivity;
-import de.danoeh.antennapod.core.storage.APCleanupAlgorithm;
-import de.danoeh.antennapod.core.storage.APNullCleanupAlgorithm;
-import de.danoeh.antennapod.core.storage.APQueueCleanupAlgorithm;
-import de.danoeh.antennapod.core.storage.EpisodeCleanupAlgorithm;
-import de.danoeh.antennapod.core.storage.EpisodeCleanupAlgorithmFactory;
-import de.danoeh.antennapod.core.storage.ExceptFavoriteCleanupAlgorithm;
+import de.danoeh.antennapod.net.download.service.episode.autodownload.APCleanupAlgorithm;
+import de.danoeh.antennapod.net.download.service.episode.autodownload.APNullCleanupAlgorithm;
+import de.danoeh.antennapod.net.download.service.episode.autodownload.APQueueCleanupAlgorithm;
+import de.danoeh.antennapod.net.download.service.episode.autodownload.EpisodeCleanupAlgorithm;
+import de.danoeh.antennapod.net.download.service.episode.autodownload.EpisodeCleanupAlgorithmFactory;
+import de.danoeh.antennapod.net.download.service.episode.autodownload.ExceptFavoriteCleanupAlgorithm;
 import de.danoeh.antennapod.storage.preferences.UserPreferences;
 import de.danoeh.antennapod.storage.preferences.UserPreferences.EnqueueLocation;
 import de.test.antennapod.EspressoTestUtils;
