+++ playback/service/build.gradle
@@ -40,5 +40,5 @@ dependencies {
     implementation "org.apache.commons:commons-lang3:$commonslangVersion"
 
     testImplementation "junit:junit:$junitVersion"
-    testImplementation 'org.mockito:mockito-core:5.11.0'
+    testImplementation "org.mockito:mockito-core:$mockitoVersion"
 }
